Update Token Validation for GitLab
As of GitLab v11.0.4 the server does _not_ sign the webhook payload with the
secret. It simply includes the secret as plain text in the X-GitLab-Token
header as a very simple validation check.

In contrast GitHub uses a more sophisticated mechanizem where the secret is
used to sign the payload and passes the sha1 in the header for validation on
the server.

This change set updates the tests and the verify function to simply ensure
that the correct header with the correct plain text secret is included
in the webhook request.