Update action type to work with GitLab
Webhook events can be fired off multiple times for a event however the object
the event is for may be moving through a work flow. This means that the same
event could be for different _actions_ being applied to the object.

In GitHub the payload had an _event_type_ modifier which we could hook up
to our events so that we only register to a particular workflow step in the
event life cycle.

GitLab has the same cycle but calls the modifier _object_attributes.action_

This change set just updates the project to use the correct property for
GitLab so that consumers can registers for given steps in the cycle vs.
all events of a given name.